Add custom UI status
This patch adds a custom UI status that can be defined and handled by
the runner.  Applications that want to use this status should have a
configuration option that lets the runner provide a value to use for
that status.  The runner is responsible for correctly handling this
status value.

Fixes https://github.com/Nitrokey/trussed/issues/5